refactor: move AgentHistoryManager to @common/history
AgentHistoryManager is a static utility class that only uses:
- @common/state/stateManager for workspace state
- @agent/core/AgentConfig for types

It has zero historyView-specific code, yet commands/agent/executeCommand.ts
was importing from @historyView/managers, creating unnecessary coupling.

Move to @common/history and re-export from historyView/managers for
backward compatibility. This eliminates the commands → historyView
dependency.
