+++ src/common/history/AgentHistoryManager.ts
