+++ src/common/history/index.ts
@@ -1,0 +1,1 @@
+export { AgentHistoryManager, type AgentHistoryItem } from './AgentHistoryManager';
